+++ app/renderer/input.js
@@ -1538,7 +1538,7 @@ const keyForOs = (regular, mac, key) => regular.includes(key)
 
 const updateNavbarScrolling = () => {
     const url = document.getElementById("url")
-    const charWidth = getSetting("fontsize") * 0.60191
+    const charWidth = getSetting("guifontsize") * 0.60191
     const end = url.selectionStart * charWidth - charWidth
     const start = url.selectionEnd * charWidth - url.clientWidth + charWidth + 2
     if (url.scrollLeft < end && url.scrollLeft > start) {